Fatal error for datagrid totals when $skipAclWalkerCheck is true
Due to introduction of the flag thr AclHelper::apply() method is not called in *all* cases. As this method internally converts QueryBuilder into Query, if it is not a Query already, the QueryBuilder does not get converted to a query if the flag is set. As there is no getScalarResult() method on QueryBuilder, this leads to a fatal error:

Fatal error: Call to undefined method Doctrine\ORM\QueryBuilder::getScalarResult() in /vagrant/vendor/oro/platform/src/Oro/Bundle/DataGridBundle/Extension/Totals/OrmTotalsExtension.php on line 308