Added attribute parameter to call_method action
This appears to have been left out not sure if it's on purpose or not but the attribute parameter is used on the call_method action in this file: https://github.com/orocrm/crm/blob/d3d34ac52808bd2778cd4b5e490618ee19c165f8/src/OroCRM/Bundle/MagentoBundle/Resources/config/process.yml

When I was reading this doc and debugging my process config, I assumed the omission of the attribute parameter meant that it was not supported in the call_method action.  After looking at the magento_customer_export action (and in oro/platform/src/Oro/Bundle/WorkflowBundle/Model/Action/CallMethod.php) , it appears that this is incorrect and that attribute parameter is supported in the call_method action.