Bugfix: Use `\DateTimeInterface` check when date object is passed
We are working with `\DateTimeInterface`, in particular with `\DateTimeImmutable` instances with Doctrine using Custom Type. Without this change, method `getDateTime()` treat our `\DateTimeImmutable` as integer and tries to pass it to `setTimestamp()` method which triggers an error.

With this update, it's possible ot use any instance of `\DateTimeInterface` interface.

Since this bundle requires PHP 7.0+, it's a safe update.

https://github.com/oroinc/platform/blob/a9b4905c85ddf2e3cebae5abc74bb9f8d6900d32/src/Oro/Bundle/LocaleBundle/composer.json#L9-L10

No BC break.