Add Webpack Progress flag to assets:build command
When using `php bin/console oro:assets:build --watch`, when files are modified nothing is displayed in the terminal output (even though the builds are running successfully). This makes it difficult to tell if Webpack has detected your code changes, or if it's actually building the assets.

Webpack has a handy `--progress` CLI parameter to watch the progress of builds:
https://webpack.js.org/api/cli/#debug-options

Adding the `--progress` parameter allows you to see a progress percentage, along with the filename currently being built.
The output looks something like this:
```
[122] ./css/layout/basetheme/styles.js 28 bytes {11} [built]
+ 221 hidden modules
65% building 220/228 modules 8 active ...public/css/layout/blank/styles.css.scss 
```

This Pull Request adds `--progress` as an optional flag, including usage instructions.