ArraySorterExtension DateTime support
We have array of objects, objects implements ArrayAccess, everything works fine except sorting by DateTime does not work. I dunno if this is the right solution, but if someone has this issue, you can extend this extension and add this. Ofc by oro logic, array should be array, but in our project we use models in some places and we need datagrid for that.